ml-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
openai==1.3.7
anthropic==0.7.7
sentence-transformers>=2.3.0
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
tiktoken==0.5.2
huggingface-hub>=0.20.0
